Improve 'gvm available' (#41)
This improves the `gvm available` command to ensure the data returned is as accurate as possible.
It addresses a few different issues.

The command uses the a cached copy of the Go git repo to list the available version tags. This will
now refresh the source repo if the cache is not from the current day.

The tags from the source repo were the only ones that were listed in the output. Each of those versions
were then annotated with whether a binary version is available. So if the source repo was stale you
would never newly released binary versions. This changes the output to merge the list of source repo
tags and binary releases so that you generally always see the latest data.

Fixes #39