adjust version naming convention for >=1.21 (#62)
This makes the assumption that starting in 1.21 all future release will always include a patch number in their naming.

For example, previous versions were named like '1.20' and then '1.20.1' (there was no '1.20.0'). Now with 1.21 the first release was versioned as '1.20.0'.

Also handle tar files without dir entries. This is also something that changed in the official Go tar distributions.